Refactored the whole repo
Moved the common structs & utils to "structs" & "utils" files
