+++ contracts/signature-bridge/src/contract.rs
@@ -80,7 +80,7 @@ fn admin_set_resource_with_signature(
     let mut data: Vec<u8> = Vec::new();
     data.extend_from_slice(&msg.resource_id);
     data.extend_from_slice(&msg.function_sig);
-    data.extend_from_slice(&element_encoder(&msg.nonce.to_le_bytes()));
+    data.extend_from_slice(&element_encoder(&msg.nonce.to_be_bytes()));
     data.extend_from_slice(&msg.new_resource_id);
     data.extend_from_slice(msg.handler_addr.as_bytes());
     data.extend_from_slice(msg.execution_context_addr.as_bytes());
