+++ packages/protocol_cosmwasm/src/lib.rs
@@ -8,7 +8,9 @@ pub mod mixer;
 pub mod mixer_verifier;
 pub mod poseidon;
 pub mod signature_bridge;
+pub mod structs;
 pub mod token_wrapper;
+pub mod utils;
 pub mod vanchor;
 pub mod vanchor_verifier;
 pub mod verifier;
